+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Changes
 - Implemented invariants for posts and profile modules (#90)
 - Added YAML support for types (#124)
+- Allow for empty message posts when they contain a poll (#148)
 
 # Version 0.4.0
 ## Changes